Pull center content out into a seperate component
